setup redis and some permissiona related work on project section

+++ src/API/project/controller/delete-a-project.controller.ts
@@ -1,15 +1,22 @@
-import { ApiErrorHandler } from "../../../shared/apiErrorHandler.js"
-import { apiResponse } from "../../../shared/apiResponseHandler.js"
-import { asyncHandler } from "../../../shared/asyncHandler.js"
-import { ProjectService } from "../service/project.service.js"
+import redis from "../../../config/redis.js";
+import { ApiErrorHandler } from "../../../shared/apiErrorHandler.js";
+import { apiResponse } from "../../../shared/apiResponseHandler.js";
+import { asyncHandler } from "../../../shared/asyncHandler.js";
+import { ProjectService } from "../service/project.service.js";
 
+export const DeleteAProjectController = asyncHandler(async (req, res) => {
+  const projectId = req.params.projectId as string;
+  if (!projectId) throw new ApiErrorHandler(404, "Project id not found");
 
-export const DeleteAProjectController = asyncHandler(async(req, res) => {
-    const projectId = req.params.projectId as string
-    if(!projectId) throw new ApiErrorHandler(404, 'Project id not found')
+  const projectService = new ProjectService();
+  const project = await projectService.findAProjectById(projectId);
+  if (!project) throw new ApiErrorHandler(404, "Project not found");
 
-    const projectService = new ProjectService()
-    await projectService.deleteAProjectById(projectId)
+  await projectService.deleteAProjectById(projectId);
 
-    res.status(200).json(new apiResponse(null, 'Success'))
-})
+  // clear cash
+  await redis.del(`all-projects:`);
+  await redis.del(`projects-by-team:${project.teamId}`);
+
+  res.status(200).json(new apiResponse(null, "Success"));
+});
